pyproject.toml
[project]
name = "awesome-cv-stats"
version = "0.1.0"
description = "Curated list of computer vision tools with automated GitHub stats"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "tenacity>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
update-stats = "scripts.update_stats:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
